push_swap : add args checker for int overflow & is_digit

+++ push_swap/libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: vmasse <vmasse@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/17 11:23:27 by vmasse            #+#    #+#             */
-/*   Updated: 2021/09/20 15:28:17 by vmasse           ###   ########.fr       */
+/*   Updated: 2021/09/20 18:13:25 by vmasse           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ static	int	check_max_long(unsigned long long int result, int sign)
 	return (-1);
 }
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	unsigned long long int		result;
 	int							sign;
@@ -44,5 +44,5 @@ int	ft_atoi(const char *str)
 	}
 	if ((result == MAX_LONG && !(*str)) || result > MAX_LONG)
 		return (check_max_long(result, sign));
-	return ((int)result * sign);
+	return (result * sign);
 }
